# spec/openapi.yaml
...
openapi: 3.1.0
info:
  title: freeq
  version: 0.1.0
  summary: AT Protocol-authenticated IRC server with a JSON API for agents.
  description: |
    freeq is an IRC server where identity is an AT Protocol DID (via the
    `ATPROTO-CHALLENGE` SASL mechanism) rather than a nickname, plus a REST
    surface so agents can read and verify conversations without speaking IRC.

    **Errors** — every `/api/v1/*` 4xx and 5xx answers with the `Error`
    schema: `error` (reason phrase, safe to branch on), `status`, `message`
    (prose, not stable), and `documentation`. A `429` carries `Retry-After`
    in seconds; a `401` carries
    `WWW-Authenticate: Bearer resource_metadata=…` (RFC 9728) naming the
    metadata document that explains how to get a token.

    **Pagination** — list endpoints that can grow use one cursor pattern:
    `limit` (bounded, server-capped at 200) plus `before`, the unix-seconds
    `timestamp` of the oldest item on the page you already have. Absent
    `before` means newest first. Paging by timestamp rather than offset means a
    page cannot shift under a concurrent write. An empty array means the end;
    there is no separate "has more" flag to disagree with the data.

    Note the cursor is the `timestamp` field, not `msgid`: msgids are ULIDs and
    sort the same way, but the server compares timestamps, so passing a msgid
    here silently returns nothing.

    **Versioning and deprecation** — the path carries the major version
...

...

...

...

...
  license:
...
servers:
...
security:
  - {}
  - bearerAuth: []
paths:
...
  /api/v1/openapi.json:
    get:
      tags: [discovery]
      operationId: getOpenApiJson
      summary: This OpenAPI document, as JSON.
      security: [{}]
      responses:
        "200":
          description: OpenAPI 3.1 document.
          content:
            application/json:
              schema:
                type: object
  /api/v1/openapi.yaml:
    get:
      tags: [discovery]
      operationId: getOpenApiYaml
      summary: This OpenAPI document, as YAML (canonical source).
      security: [{}]
      responses:
        "200":
          description: OpenAPI 3.1 document.
          content:
            application/yaml:
              schema:
                type: string
  /.well-known/agent.json:
    get:
      tags: [discovery, agent-assistance]
      operationId: getAgentDiscovery
      summary: Agent Assistance Interface discovery document.
      description: |
        Lists the assistance tools this server actually has mounted.
        `capabilities` is kept in lock-step with the `/agent/tools/*` routes;
        `free_form_session` only appears when an LLM provider is configured.
      security: [{}]
      responses:
        "200":
          description: Discovery document.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentDiscovery"
  /llms.txt:
    get:
      tags: [discovery]
      operationId: getLlmsTxt
      summary: Markdown index of this server's agent surfaces.
      security: [{}]
      responses:
        "200":
          description: llms.txt document.
          content:
            text/markdown:
              schema:
                type: string
...
  /openapi.json:
    get:
      tags: [discovery]
      operationId: getOpenapiAtRoot
      summary: Alias for /api/v1/openapi.json, where crawlers look first.
      security: [{}]
      responses:
        "200":
          description: This document.
          content:
            application/json:
              schema:
                type: object
...
  /api/v1/channels:
    get:
      tags: [channels]
      operationId: listChannels
      summary: List visible channels.
      security: [{}]
      responses:
        "200":
          description: Channels with member counts and topics.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChannelInfo"
...
  /api/v1/verify/{msgid}:
    get:
      tags: [messages]
      operationId: verifyMessage
      summary: Verify a message's signature.
      description: |
        Re-canonicalizes the stored message, resolves the signing key (client
        session key or server key) and reports whether the signature holds.
      parameters:
        - $ref: "#/components/parameters/MsgId"
      security: [{}]
      responses:
        "200":
          description: Verification result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyResult"
        "404": { $ref: "#/components/responses/NotFound" }

  # ── users ────────────────────────────────────────────────────────────
...
  /agent/session:
    post:
      tags: [agent-assistance]
      operationId: agentSession
      summary: Free-form assistance request, routed to a tool by an LLM.
      description: |
        Returns `LLM_NOT_CONFIGURED` when no provider is installed; in that
        case call a structured `/agent/tools/*` endpoint directly.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message: { type: string }
                context:
                  type: object
                  description: Reserved for future structured context.
      responses:
        "200": { $ref: "#/components/responses/AgentTool" }

  # ── coordination ─────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        A freeq API bearer token. Obtained either from the `API-BEARER`
        NOTICE the server sends after a successful SASL `903`, or from
        `POST /auth/broker/web-token`. Identifies the caller by DID.
    brokerSignature:
      type: apiKey
      in: header
      name: X-Broker-Signature
      description: HMAC over the raw request body, using the broker shared secret.
  parameters:
...
    MsgId:
      name: msgid
      in: path
      required: true
      description: ULID message id.
      schema: { type: string }
...
  responses:
    AgentTool:
      description: |
        Evidence bundle envelope: a conclusion plus the evidence behind it,
        with a `request_id` for audit correlation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/AgentToolEnvelope"
...
    NotFound:
      description: No such resource.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      description: |
        The error envelope for every `/api/v1/*` response with a 4xx or 5xx
        status. A middleware normalises errors that handlers (or axum itself,
        for an unrouted path or an unsupported method) would otherwise return
        as a bare string, so a client can parse one shape rather than three.

        `error` is the HTTP reason phrase and is stable enough to branch on;
        `message` is human prose and is not. Errors outside `/api/` — the
        `/verify/*` pages and the markdown 404 — are for people and crawlers
        and stay HTML or markdown.
      required: [error, status, message]
      properties:
        error:
          type: string
          description: HTTP reason phrase, e.g. `Not Found`.
          examples: ["Not Found", "Unauthorized", "Too Many Requests"]
        status:
          type: integer
          description: The HTTP status code, repeated in the body.
          examples: [404]
        message:
          type: string
          description: What went wrong, in prose. Do not branch on this.
        documentation:
          type: string
          format: uri
          description: Where the contract for this endpoint lives.
      examples:
        - error: Not Found
          status: 404
          message: "No such endpoint: /api/v1/definitely-not-a-route"
          documentation: https://irc.freeq.at/api/v1/openapi.json
...
    ChannelInfo:
      type: object
      required: [name, members]
      properties:
        name: { type: string }
        members: { type: integer }
        topic: { type: [string, "null"] }
...
    VerifyResult:
      type: object
      properties:
        msgid: { type: string }
        verified: { type: boolean }
        signer: { type: string, description: DID or server name. }
        signed_by:
          type: string
          description: '`client` (session key) or `server` (fallback).'
        reason:
          type: string
          description: Why verification failed, when it did.
...
    AgentDiscovery:
      type: object
      required: [service, version, assistance_endpoint, capabilities, auth]
      properties:
        service: { type: string, examples: [Freeq] }
        version: { type: string }
        description: { type: string }
        assistance_endpoint: { type: string, examples: ["/agent/tools"] }
        capabilities:
          type: array
          items: { type: string }
        auth:
          type: object
          properties:
            required: { type: boolean }
            methods:
              type: array
              items: { type: string }
    AgentToolEnvelope:
      type: object
      properties:
        request_id: { type: string }
        conclusion: { type: string }
        evidence: { type: object }
        caller: { type: object }